move to `@sentry/integrations`
