+++ packages/browser-integration-tests/suites/integrations/ContextLines/init.js
@@ -1,8 +1,9 @@
 import * as Sentry from '@sentry/browser';
+import { ContextLines } from '@sentry/integrations';
 
 window.Sentry = Sentry;
 
 Sentry.init({
   dsn: 'https://public@dsn.ingest.sentry.io/1337',
-  integrations: [new Sentry.Integrations.ContextLines()],
+  integrations: [new ContextLines()],
 });
