+++ packages/browser/src/integrations/index.ts
@@ -4,4 +4,3 @@ export { Breadcrumbs } from './breadcrumbs';
 export { LinkedErrors } from './linkederrors';
 export { HttpContext } from './httpcontext';
 export { Dedupe } from './dedupe';
-export { ContextLines } from './contextlines';
