+++ packages/integrations/src/index.ts
@@ -9,3 +9,4 @@ export { RewriteFrames } from './rewriteframes';
 export { SessionTiming } from './sessiontiming';
 export { Transaction } from './transaction';
 export { HttpClient } from './httpclient';
+export { ContextLines } from './contextlines';
