+++ packages/integrations/test/contextlines.test.ts
@@ -1,6 +1,6 @@
 import type { StackFrame } from '@sentry/types';
 
-import { applySourceContextToFrame } from '../../../src/integrations/contextlines';
+import { applySourceContextToFrame } from '../src/contextlines';
 
 const lines = ['line1', 'line2', 'line3', 'line4', 'line5', 'line6', 'line7', 'line8', 'line9'];
 describe('ContextLines', () => {
